Try to fix all MSVC warnings (W4)

+++ include/cat_memory.h
@@ -92,7 +92,7 @@ CAT_API void cat_freep_function(void *ptr); /* free(ptr->ptr) */
 CAT_API char *cat_strdup_function(const char *string);
 CAT_API char *cat_strndup_function(const char *string, size_t length);
 
-CAT_API size_t cat_getpagesize(void);
+CAT_API unsigned int cat_getpagesize(void);
 CAT_API void *cat_getpageof(const void *p);
 
 CAT_API unsigned int cat_bit_count(uintmax_t num);
